No need for bokehjs/ts-node anymore
