--- bokehjs/ts-node/register.js
@@ -1,1 +1,0 @@
-require("./").register({cache: false})
